+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Rewritten in Go
 - Work with OpenSearch v3.2 as a backend instead of Elasticsearch v6.8
+- Let searches return 10 items instead of 30 by default
 
 ### Removed